requirements.txt
flake8>=2.1.0
pytest-cov>=1.6
